Add logic for missing Start Times for Trips
Some older datasets do not provide start times for vehicle positions and
trip updates. For those datasets, we need to pull start times from the
static schedule (for planned trips) or compute it from the data we do
have (added trips which are, by definition, not in the static schedule).

* Add a function to `l1_rt_trips.py` that adds the start time to the
  TempVehicleTrips table.
* Call that function when running `proccess_trips`.